experiment/3_products/requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
mysql-connector-python>=8.0.33
lxml>=4.9.0

